submit quiz

+++ database/migrations/2024_09_27_013748_create_user_quizzes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->foreignUuid('quiz_id')->constrained()->cascadeOnDelete()->cascadeOnUpdate();
             $table->text('module_question_id');
             $table->text('answer')->nullable();
-            $table->integer('score')->default(0);
+            $table->double('score')->default(0);
             $table->timestamps();
         });
     }
